# src/main/java/io/swagger/api/SendAlertEventApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.SendAlertEventPOSTRequest;
import io.swagger.model.SendAlertEventPOSTResponse;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-02-19T13:22:33.560Z")

@Api(value = "/sendAlertEvent", description = "the /sendAlertEvent API")
public interface SendAlertEventApi {

    @ApiOperation(value = "SendAlertEvent_POST", nickname = "sendAlertEventPost", notes = "", response = SendAlertEventPOSTResponse.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "", response = SendAlertEventPOSTResponse.class) })
    @RequestMapping(value = "//sendAlertEvent",
        produces = { "application/json" },
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<SendAlertEventPOSTResponse> sendAlertEventPost(@ApiParam(value = "" ,required=true )  @Valid @RequestBody SendAlertEventPOSTRequest body);

}